requirements.txt
fastapi==0.115.6
uvicorn==0.34.0
langchain==0.3.13
langchain-community==0.3.12
langchain-core==0.3.27
langchain-cohere==0.3.3
cohere==5.13.3
openai==1.58.1
langchain-openai==0.2.13
python-multipart==0.0.20
pypdf==5.1.0
beautifulsoup4==4.12.3
requests==2.32.3
python-dotenv==1.0.1
sse-starlette==2.1.3
cassio==0.1.3
pydantic==2.10.4
astrapy==1.5.2
